helmet/ml/models/model_optimiser.py
import torch


def model_optimiser(model):
        params = [p for p in model.parameters() if p.requires_grad]
        optimizer = torch.optim.SGD(params, lr=0.01, momentum=0.9, nesterov=True, weight_decay=1e-4)
        return optimizer
